不要なプロジェクト関連のクエリファイルを削除し、プロジェクトリストおよび管理タブのインポートパスを修正しました。また、プロジェクト詳細コンポーネントを動的インポートに変更しました。
