+++ next/src/app/[locale]/(common-layout)/user/[handle]/_components/user-project-list/server.tsx
@@ -1,4 +1,4 @@
-import { fetchProjectsWithPagination } from "./_db/queries.server";
+import { fetchProjectsWithPagination } from "../../_db/queries.server";
 import { UserProjectListClient } from "./client";
 
 interface UserProjectListProps {
