+++ next/src/app/[locale]/(common-layout)/user/[handle]/_db/queries.server.ts
@@ -63,3 +63,55 @@ export async function fetchFollowingList(userId: string) {
 		following: sanitizeUser(record.following),
 	}));
 }
+
+export async function fetchProjectsWithPagination(
+	userId: string,
+	page = 1,
+	pageSize = 10,
+	searchTerm = "",
+) {
+	const skip = (page - 1) * pageSize;
+	const whereClause = {
+		userId,
+		OR: [
+			{
+				title: {
+					contains: searchTerm,
+					mode: "insensitive" as const,
+				},
+			},
+			{
+				description: {
+					contains: searchTerm,
+					mode: "insensitive" as const,
+				},
+			},
+		],
+	};
+
+	const [projectsWithRelations, totalCount] = await Promise.all([
+		prisma.project.findMany({
+			where: whereClause,
+			orderBy: {
+				updatedAt: "desc",
+			},
+			skip,
+			take: pageSize,
+			include: {
+				user: true,
+				images: true,
+				links: true,
+				projectTagRelations: { include: { projectTag: true } },
+			},
+		}),
+		prisma.project.count({
+			where: whereClause,
+		}),
+	]);
+
+	return {
+		projectsWithRelations,
+		totalPages: Math.ceil(totalCount / pageSize),
+		currentPage: page,
+	};
+}
