+++ next/src/app/[locale]/(common-layout)/user/[handle]/project-management/_components/project-management-tab/server.tsx
@@ -1,4 +1,4 @@
-import { fetchProjectsWithPagination } from "../../_db/queries.server";
+import { fetchProjectsWithPagination } from "../../../_db/queries.server";
 import { ProjectManagementTabClient } from "./client";
 
 interface ProjectManagementTabProps {
